+++ nautilus_trader/data/aggregation.pyx
@@ -329,9 +329,9 @@ cdef class TickBarAggregator(BarAggregator):
 
         """
         super().__init__(
-            bar_type,
-            handler,
-            logger,
+            bar_type=bar_type,
+            handler=handler,
+            logger=logger,
             use_previous_close=False,
         )
 
@@ -370,9 +370,9 @@ cdef class VolumeBarAggregator(BarAggregator):
 
         """
         super().__init__(
-            bar_type,
-            handler,
-            logger,
+            bar_type=bar_type,
+            handler=handler,
+            logger=logger,
             use_previous_close=False,
         )
 
@@ -433,9 +433,9 @@ cdef class ValueBarAggregator(BarAggregator):
 
         """
         super().__init__(
-            bar_type,
-            handler,
-            logger,
+            bar_type=bar_type,
+            handler=handler,
+            logger=logger,
             use_previous_close=False,
         )
 
@@ -518,9 +518,9 @@ cdef class TimeBarAggregator(BarAggregator):
 
         """
         super().__init__(
-            bar_type,
-            handler,
-            logger,
+            bar_type=bar_type,
+            handler=handler,
+            logger=logger,
             use_previous_close=use_previous_close,
         )