+++ nautilus_trader/model/instruments/currency.pyx
@@ -148,7 +148,6 @@ cdef class CurrencySpot(Instrument):
             asset_class=asset_class,
             asset_type=AssetType.SPOT,
             quote_currency=quote_currency,
-            cost_currency=quote_currency,
             is_inverse=False,
             price_precision=price_precision,
             size_precision=size_precision,