+++ tests/integration_tests/adapters/betfair/conftest.py
@@ -26,6 +26,7 @@ from nautilus_trader.common.clock import LiveClock
 from nautilus_trader.common.enums import LogLevel
 from nautilus_trader.common.logging import LiveLogger
 from nautilus_trader.common.uuid import UUIDFactory
+from nautilus_trader.model.currencies import AUD
 from nautilus_trader.model.identifiers import InstrumentId
 from nautilus_trader.model.identifiers import PositionId
 from nautilus_trader.model.identifiers import Symbol
@@ -198,6 +199,7 @@ async def execution_client(
     client = BetfairExecutionClient(
         client=betfair_client,
         account_id=account_id,
+        base_currency=AUD,
         engine=exec_engine,
         clock=clock,
         logger=live_logger,