+++ tests/unit_tests/live/test_live_execution_client.py
@@ -24,6 +24,8 @@ from nautilus_trader.execution.messages import OrderStatusReport
 from nautilus_trader.live.execution_engine import LiveExecutionEngine
 from nautilus_trader.live.risk_engine import LiveRiskEngine
 from nautilus_trader.model.commands import SubmitOrder
+from nautilus_trader.model.currencies import USD
+from nautilus_trader.model.enums import AccountType
 from nautilus_trader.model.enums import OrderSide
 from nautilus_trader.model.enums import OrderState
 from nautilus_trader.model.enums import VenueType
@@ -60,7 +62,6 @@ class TestLiveExecutionClient:
         self.logger = Logger(self.clock)
 
         self.trader_id = TraderId("TESTER-000")
-        self.account_id = TestStubs.account_id()
 
         self.order_factory = OrderFactory(
             trader_id=self.trader_id,
@@ -100,7 +101,9 @@ class TestLiveExecutionClient:
         self.client = MockLiveExecutionClient(
             client_id=ClientId(SIM.value),
             venue_type=VenueType.ECN,
-            account_id=self.account_id,
+            account_id=TestStubs.account_id(),
+            account_type=AccountType.CASH,
+            base_currency=USD,
             engine=self.exec_engine,
             instrument_provider=InstrumentProvider(),
             clock=self.clock,